// client/src/actions/activity/ActivityActionTypes.enum.ts
enum ActivityActionTypes {
  GET_ACTIVITIES = "GET_ACTIVITIES",
  SET_CURRENT_ACTIVITY = "SET_CURRENT_ACTIVITY",
  CLEAR_CURRENT_ACTIVITY = "CLEAR_CURRENT_ACTIVITY",
  SET_CREATE_MODE = "SET_CREATE_MODE",
  SET_EDIT_MODE = "SET_EDIT_MODE",
  OPEN_CREATE_FORM = "OPEN_CREATE_FORM",
  OPEN_EDIT_FORM = "OPEN_EDIT_FORM",
  CREATE_ACTIVITY = "CREATE_ACTIVITY",
  UPDATE_ACTIVITY = "UPDATE_ACTIVITY",
  DELETE_ACTIVITY = "DELETE_ACTIVITY"
}

export default ActivityActionTypes;